refactor: move Cart Dropdown to own component
